fixed bugs

+++ app/views/classes/index.php
@@ -29,19 +29,16 @@
     </thead>
     <tbody>
         <?php foreach ($data["Class"] as $class) { ?>
-        <tr>
+            <tr>
 
-            <td><a href="<?= URLROOT; ?>studentsController/index/<?= $class->classId ?>"
-                    style="text-decoration:none; color:black;"><?= $class->classId ?></a></td>
-            <td><?= $class->classTeacherId ?></td>
-            <td><?= $class->className ?></td>
-            <td style="align-items: center;">
-                <a class=" btn btn-danger" href="<?= URLROOT; ?>/classesController/delete/<?= $class->classId ?>"
-                    onclick="return confirm('Are you sure you want to delete this school?')">Delete</a>
-                <a class="btn btn-info"
-                    href="<?= URLROOT; ?>/classesController/update/<?= $class->classId ?>">Update</a>
-            </td>
-        </tr>
+                <td><a href="<?= URLROOT; ?>studentsController/index/<?= $class->classId ?>" style="text-decoration:none; color:black;"><?= $class->classId ?></a></td>
+                <td><?= $class->classTeacherId ?></td>
+                <td><?= $class->className ?></td>
+                <td style="align-items: center;">
+                    <a class=" btn btn-danger" href="<?= URLROOT; ?>/classesController/delete/<?= $class->classId . '+' . $data['Teacher']->teacherId  ?>" onclick="return confirm('Are you sure you want to delete this school?')">Delete</a>
+                    <a class="btn btn-info" href="<?= URLROOT; ?>/classesController/update/<?= $class->classId . '+' . $data['Teacher']->teacherId ?>">Update</a>
+                </td>
+            </tr>
         <?php } ?>
     </tbody>
 </table>
